services/datahub-obsidian-source/pyproject.toml
[project]
name = "datahub-obsidian-source"
version = "0.1.0"
description = "DataHub ingestion source for Obsidian vault markdown files"
requires-python = ">=3.12"
dependencies = [
    "acryl-datahub>=0.14.0",
    "pyyaml>=6.0",
]

[project.entry-points."datahub.ingestion.source.plugins"]
obsidian = "obsidian_source.source:ObsidianSource"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/obsidian_source"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["pytest>=8.0", "ruff>=0.9"]
